backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
env==0.1.0
executing==2.0.1
filelock==3.14.0
icecream==2.1.3
idna==3.10
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.0
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
pyscript==0.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
pyzmq==26.0.3
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
toml==0.10.2
tornado==6.4
traitlets==5.14.3
typer==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.2
wcwidth==0.2.13
